Correctly handle Development item in popup 'global menu'.
(semantic-mode-map): To avoid showing showing Development menu twice,
only disable menu item if menu-bar is actually enabled, otherwise the
popup 'global menu' might display a disabled Development menu.

This seems to be an inconsistency in how the menu-bar and the 'global
menu' popup is handled. The latter never shows the menu item twice if
both entries from Semantic and EDE are enabled, contrary to the
menu-bar. If one of them is disabled, the global menu shows one of
them (mostly the disabled one, but it seems to be racy).