Fix problems with Emacs lisp database.
* semantic/db-el.el (semanticdb-normalize-one-tag): It might be that a
symbol comes from a file but cannot be found in its table.  This
happens for instance when a symbol was dynamically created through a
macro like `defstruct'.  In this case, return the original tag.
(semanticdb-elisp-sym->tag): Deal with autoloaded functions, where the
argument list is not available until the file is loaded.