Also get local variables for unparseable buffers (revert rev. 4997)
(semantic-get-local-variables-default): Also try to parse local
variables for buffers which are currently marked as unparseable.
Otherwise, it is often impossible to complete local variables, for
example when the closing paren in an 'if' condition is missing.  While
this might make the function throw an error, we should rather fix this
through `condition-case'.